Update and rename ramachandran_hw3.py to Differentiation.py
Differentials of the Numbers of Early Universe Starbursts Submillimeter-selected galaxies are the
brightest galaxies in the Universe. They live roughly 1-2 billion years after the Big Bang (in our
14 billion year Universe). Since we’ve known about them, in 1998, theorists have been scratching
their heads trying to find a model to match their huge abundances. In short, when we look in our
simulations, we have a really hard time finding as many galaxies that are as bright as observers
observe.
Imagine you have a model where the data are given on the class bitbucket repository. The model data are called ’model_smg.dat’.
These model data are created as the number of galaxies above a given luminosity. The first column is the
luminosity (in a unit we’ll call ’millijansky’), and the second column is the number of galaxies above
that luminosity (N > L).
At the same time, there are observed data from various surveys in a file called ’ncounts_850.dat’. The
first column is the logarithm of the luminosity, while the second column is the logartihm of dN/dL.
Write a code to convert the absolute model number counts to differential counts, and plot the observed
data against the model.