Update and rename ramachandran_hw4.py to Integration.py
Integration – How Big are Objects in the Universe? In cosmology, the angular diameter distance
defines the distance of an object in terms of the angular size of the object as viewed from earth. In
other words, the size of a distance on the sky is dependent on it’s distance. You can imagine some
toy scenario where your friend is walking away from you in some big flat field. The further they get
from you, the smaller they will look. We say that this person subtends some angle, θ.
Nominally, for small angles, we know that θ ≈ l/D, where l is the size of the object, and D is the
distance to the object. In cosmology, however, things are crazy. The curvature of the Universe is set
by a competition of the expansion rate, and gravity pulling objects closer together.
The angular diameter distance in cosmology, DA can be described as:
DA = (intergration from z2 to z1 of r(z) dz)/(1+z)
where r(z) = C/H(z), and H(z) = sqrt(ΩM(1 + z)^3 + ΩR(1 + z)^2+ ΩΛ), and z1 and z2 are the redshift
range you’re considering. (Typically z1 = 0). C = 3 × 10^3, Ωm = 0.3, ΩΛ = 0.7 and ΩR = 0.
Here, I want you to calculate DA for a range of redshifts (from z1 = 0), with a maximum redshift of
z2 = 5). To integrate the function, write a trapezoidal rule integrator. You will build up the function
in pieces integrating (for example) from z1 = 0 to z2 = 1, then from z1 = 0 to z1 = 3 etc